Separating the plugin framework to a separate class
 - a class with generic type of the loaded plugins
 - making the authorization framework a subclass of it
approaches #2588
