+++ opengrok-indexer/src/main/java/org/opengrok/indexer/framework/PluginClassLoader.java
@@ -17,10 +17,10 @@
  * CDDL HEADER END
  */
 
- /*
+/*
  * Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
  */
-package org.opengrok.indexer.authorization;
+package org.opengrok.indexer.framework;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -37,39 +37,39 @@ import java.util.logging.Logger;
 import org.opengrok.indexer.logger.LoggerFactory;
 
 /**
- * Class loader for authorization plugins.
+ * Class loader for plugins from .class and .jar files.
  *
  * @author Krystof Tulinger
  */
-public class AuthorizationPluginClassLoader extends ClassLoader {
+public class PluginClassLoader extends ClassLoader {
 
     @SuppressWarnings("rawtypes")
     private final Map<String, Class> cache = new HashMap<>();
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(AuthorizationPluginClassLoader.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(PluginClassLoader.class);
     private static final String[] CLASS_WHITELIST = new String[]{
-        "org.opengrok.indexer.configuration.Group",
-        "org.opengrok.indexer.configuration.Project",
-        "org.opengrok.indexer.configuration.RuntimeEnvironment",
-        "org.opengrok.indexer.authorization.IAuthorizationPlugin",
-        "org.opengrok.indexer.authorization.plugins.*",
-        "org.opengrok.indexer.util.*",
-        "org.opengrok.indexer.logger.*"
+            "org.opengrok.indexer.configuration.Group",
+            "org.opengrok.indexer.configuration.Project",
+            "org.opengrok.indexer.configuration.RuntimeEnvironment",
+            "org.opengrok.indexer.authorization.IAuthorizationPlugin",
+            "org.opengrok.indexer.authorization.plugins.*",
+            "org.opengrok.indexer.util.*",
+            "org.opengrok.indexer.logger.*"
     };
 
     private static final String[] PACKAGE_BLACKLIST = new String[]{
-        "java",
-        "javax",
-        "org.w3c",
-        "org.xml",
-        "org.omg",
-        "sun"
+            "java",
+            "javax",
+            "org.w3c",
+            "org.xml",
+            "org.omg",
+            "sun"
     };
 
     private final File directory;
 
-    public AuthorizationPluginClassLoader(File directory) {
-        super(AuthorizationPluginClassLoader.class.getClassLoader());
+    public PluginClassLoader(File directory) {
+        super(PluginClassLoader.class.getClassLoader());
         this.directory = directory;
     }
 
@@ -99,8 +99,8 @@ public class AuthorizationPluginClassLoader extends ClassLoader {
                         Class c = defineClass(classname, bytes, 0, bytes.length);
                         LOGGER.log(Level.FINE, "Class \"{0}\" found in file \"{1}\"",
                                 new Object[]{
-                                    classname,
-                                    f.getAbsolutePath()
+                                        classname,
+                                        f.getAbsolutePath()
                                 });
                         return c;
                     }
@@ -125,8 +125,8 @@ public class AuthorizationPluginClassLoader extends ClassLoader {
                 Class c = defineClass(classname, bytes, 0, bytes.length);
                 LOGGER.log(Level.FINEST, "Class \"{0}\" found in file \"{1}\"",
                         new Object[]{
-                            classname,
-                            f.getAbsolutePath()
+                                classname,
+                                f.getAbsolutePath()
                         });
                 return c;
             }
@@ -178,7 +178,7 @@ public class AuthorizationPluginClassLoader extends ClassLoader {
 
     /**
      * Loads the class with given name.
-     *
+     * <p>
      * Order of lookup:
      * <ol>
      * <li>already loaded classes </li>
@@ -186,14 +186,14 @@ public class AuthorizationPluginClassLoader extends ClassLoader {
      * <li>loading from .class files</li>
      * <li>loading from .jar files</li>
      * </ol>
-     *
+     * <p>
      * Package blacklist: {@link #PACKAGE_BLACKLIST}.<br>
      * Classes whitelist: {@link #CLASS_WHITELIST}.
      *
      * @param name class name
      * @return loaded class or null
      * @throws ClassNotFoundException if class is not found
-     * @throws SecurityException if the loader cannot access the class
+     * @throws SecurityException      if the loader cannot access the class
      */
     @Override
     public Class<?> loadClass(String name) throws ClassNotFoundException, SecurityException {
@@ -202,7 +202,7 @@ public class AuthorizationPluginClassLoader extends ClassLoader {
 
     /**
      * Loads the class with given name.
-     *
+     * <p>
      * Order of lookup:
      * <ol>
      * <li>already loaded classes </li>
@@ -210,15 +210,15 @@ public class AuthorizationPluginClassLoader extends ClassLoader {
      * <li>loading from .class files</li>
      * <li>loading from .jar files</li>
      * </ol>
-     *
+     * <p>
      * Package blacklist: {@link #PACKAGE_BLACKLIST}.<br>
      * Classes whitelist: {@link #CLASS_WHITELIST}.
      *
-     * @param name class name
+     * @param name      class name
      * @param resolveIt if the class should be resolved
      * @return loaded class or null
      * @throws ClassNotFoundException if class is not found
-     * @throws SecurityException if the loader cannot access the class
+     * @throws SecurityException      if the loader cannot access the class
      */
     @Override
     @SuppressWarnings("rawtypes")
